WebSite/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alibabacloud-credentials==1.0.2
alibabacloud-credentials-api==1.0.0
alibabacloud-dysmsapi20170525==4.1.2
alibabacloud-tea==0.4.3
alibabacloud_endpoint_util==0.0.4
alibabacloud_gateway_spi==0.0.3
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_openapi==0.3.16
alibabacloud_tea_util==0.3.13
alibabacloud_tea_xml==0.0.3
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-dysmsapi==2.1.2
aliyunsdkcore==1.0.3
APScheduler==3.11.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
Django==5.2.4
et_xmlfile==2.0.0
exceptiongroup==1.3.0
frozenlist==1.7.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==0.10.0
multidict==6.6.3
mysqlclient==2.1.1
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
priority==2.0.0
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycrypto==2.6.1
pycryptodome==3.23.0
pycryptodomex==3.23.0
PyMySQL==1.1.1
pyOpenSSL==25.1.0
python-alipay-sdk==3.3.0
pytz==2025.2
reportlab==4.4.2
requests==2.32.4
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
taskgroup==0.2.2
tomli==2.2.1
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
uWSGI==2.0.30
wsproto==1.2.0
yarl==1.20.1
zope.interface==7.2
